+++ src/main/kotlin/dev/micropythontools/ui/MpyActions.kt
@@ -46,6 +46,8 @@ import com.intellij.testFramework.LightVirtualFile
 import com.intellij.util.PathUtilRt
 import com.intellij.util.PathUtilRt.Platform
 import com.intellij.util.asSafely
+import com.intellij.util.ui.UIUtil
+import com.jediterm.terminal.ui.JediTermWidget
 import com.jetbrains.python.PythonFileType
 import dev.micropythontools.communication.MpyDeviceService
 import dev.micropythontools.communication.MpyTransferService
@@ -202,8 +204,8 @@ internal abstract class MpyUploadActionBase(
         visibleWhen = VisibleWhen.PLUGIN_ENABLED,
         enabledWhen = EnabledWhen.PLUGIN_ENABLED,
         requiresConnection = true,
-        requiresRefreshAfter = true,
-        cancelledMessage = "Upload operation cancelled"
+        requiresRefreshAfter = false, // A manual performReplAction is called by the uploading class
+        cancelledMessage = "Upload cancelled"
     )
 ) {
     init {
@@ -264,7 +266,7 @@ internal class MpyRefreshAction : MpyReplAction(
         visibleWhen = VisibleWhen.PLUGIN_ENABLED,
         enabledWhen = EnabledWhen.CONNECTED,
         requiresConnection = false,
-        requiresRefreshAfter = false,
+        requiresRefreshAfter = false, // This option isn't used, refresh is instead called explicitly, (allows cancellation)
         cancelledMessage = "Refresh operation cancelled"
     )
 ) {
@@ -363,6 +365,10 @@ internal class MpyDeleteAction : MpyReplAction(
             .toSet()
 
         deviceService.recursivelySafeDeletePaths(pathsToDelete)
+    }
+
+    override fun customUpdate(e: AnActionEvent) {
+        e.presentation.isEnabled = !deviceService.fileSystemWidget?.selectedFiles().isNullOrEmpty()
     }
 
     override fun dialogToShowFirst(e: AnActionEvent): DialogResult {
@@ -600,6 +606,29 @@ internal class MpySoftResetAction : MpyReplAction(
     override suspend fun performAction(e: AnActionEvent, reporter: RawProgressReporter) {
         reporter.text("Resetting...")
         deviceService.reset()
+    }
+}
+
+internal class MpyClearReplAction : MpyAction(
+    "Clear REPL",
+    MpyActionOptions(
+        visibleWhen = VisibleWhen.ALWAYS,
+        enabledWhen = EnabledWhen.ALWAYS,
+        requiresConnection = false,
+        requiresRefreshAfter = false
+    )
+) {
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
+
+    override fun performAction(e: AnActionEvent) {
+        val sure =
+            MessageDialogBuilder.yesNo("Clear REPL", "Are you sure you want to clear the REPL terminal?").ask(project)
+
+        if (sure) {
+            val terminal = project.service<MpyComponentRegistryService>().getTerminal()
+            val widget = UIUtil.findComponentOfType(terminal?.component, JediTermWidget::class.java)
+            widget?.terminalPanel?.clearBuffer()
+        }
     }
 }